+++ lib/hexapdf/stream.rb
@@ -181,17 +181,16 @@ module HexaPDF
       source
     end
 
+    # :call-seq:
+    #   stream.stream_encoder
+    #
     # Returns the encoder Fiber for the stream data.
     #
-    # The two arguments can be used to add additional filters for *only* this returned encoder
-    # Fiber. They should normally *not* be used and are here for use by the encryption facilities.
-    #
     # See the Filter module for more information on how to work with the fiber.
-    def stream_encoder(additional_filter = nil, additional_decode_parms = nil)
-      encoder_data = [additional_filter, document.unwrap(self[:Filter])].flatten.
-        zip([additional_decode_parms, document.unwrap(self[:DecodeParms])].flatten).
+    def stream_encoder(source = stream_source)
+      encoder_data = [document.unwrap(self[:Filter])].flatten.
+        zip([document.unwrap(self[:DecodeParms])].flatten).
         delete_if {|f, _| f.nil?}
-      source = stream_source
 
       if data.stream.kind_of?(StreamData)
         decoder_data = data.stream.filter.zip(data.stream.decode_parms)